web/requirements.txt
flask
werkzeug
python-dotenv
psycopg2
sqlalchemy
sqlalchemy-serializer
waitress